+++ standingssync/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "standingssync.apps.SyncAltContactsConfig"
 
-__version__ = "1.2.0b2"
+__version__ = "1.2.0b3"
 __title__ = "Standings Sync"